pyproject.toml
[tool.poetry]
name = "oh_sched_web"
version = "0.1.0"
description = ""
authors = ["matt higger <matt.higger@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.1.0"
pyyaml = "^6.0.2"
gunicorn = "^23.0.0"
boto3 = "^1.38.3"
python-dotenv = "^1.1.0"
oh-sched = "^0.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
